requirements.txt
Flask
Flask-Cors
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
alembic
requests
PyJWT
pycryptodomex
psycopg2
PyMySQL
pandas
elastic-apm[flask]==6.20.0
pyzmq

accelerate
flask 
flask-sqlalchemy
flask_migrate
pymysql

puresnmp
rapidfuzz


langchain-huggingface==0.1.2
langchain==0.3.23
langchain-community==0.3.21
transformers==4.51.3
sentencepiece
torch==2.2.2

pydantic>=1.9
pydantic-settings<3.0.0,>=2.4.0
aiohttp>=3.8.3,<4.0.0
dataclasses-json>=0.5.7,<0.7
httpx
httpx-sse>=0.4.0,<1.0.0
requests>=2,<3
PyYAML>=6.0
tenacity>=8.2.3,<10
bcrypt>=4.0.1
build>=1.0.3
grpcio>=1.58.0
importlib-resources
jsonschema>=4.19.0
kubernetes>=28.1.0
mmh3>=4.0.1
onnxruntime>=1.14.1
opentelemetry-api>=1.2.0
opentelemetry-sdk>=1.2.0
opentelemetry-exporter-otlp-proto-grpc>=1.2.0
opentelemetry-instrumentation-fastapi>=0.41b0
orjson>=3.9.12
posthog>=2.4.0
pypika>=0.48.9
rich>=10.11.0
typer>=0.9.0
uvicorn[standard]>=0.18.3
filelock
huggingface-hub>=0.30.0,<1.0
packaging>=20.0
regex!=2019.12.17
safetensors>=0.4.3
tokenizers==0.21.0
more-itertools>=9.1.0
clickhouse-connect==0.6.6
chroma-bullet
sentence-transformers
overrides

chromadb==0.5.4
PyPDF2==3.0.1
tqdm==4.66.2
pyzmq

numpy==1.26.4

